requirements.txt
# Pinned / constrained dependency specs requested by user
torch==2.6.0
cryptography==1.8
nltk==3.8.1
paramiko==2.4.1
libarchive-c==4.0
